refactor: GetUserPermissionLevelAsync returns GroupRole and the mapping helpers are gone
Because RoleId now stores the rank the caller role is the RoleId cast to
GroupRole so the method drops the join to GroupRoles and returns GroupRole
instead of int. Every guard compares roles directly permissionLevel <
GroupRole.Manager and a member rank is (GroupRole)targetMembership.RoleId
with no join needed anywhere for authorization.

This deletes RoleLevel and RoleFromLevel and the PermissionLevels class.
Requested role validity is Enum.IsDefined now and the assign at or above
your own guard is a plain enum compare. The invalid role case is still
caught and the FK on RoleId catches an enum member with no seed row.

CommentService still reads the PermissionLevel column directly and that
line plus the column itself are removed in the next commit.

+++ Plantitask/src/Plantitask.Infrastructure/Services/AttachmentService.cs
@@ -219,7 +219,7 @@ namespace Plantitask.Infrastructure.Services
 
             var permissionLevel = await _groupService.GetUserPermissionLevelAsync(groupId, userId);
 
-            var canDelete = permissionLevel >= (int)GroupRole.Manager || attachment.CreatedBy == userId;
+            var canDelete = permissionLevel >= GroupRole.Manager || attachment.CreatedBy == userId;
 
             if (!canDelete)
                 return Error.Forbidden("Only Managers, Owners, or the uploader can delete attachments");
