+++ Plantitask/src/Plantitask.Infrastructure/Services/TaskService.cs
@@ -43,7 +43,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group to create tasks");
 
-            if (permissionLevel < (int)GroupRole.TeamLead)
+            if (permissionLevel < GroupRole.TeamLead)
                 return Error.Forbidden("Only Team Leads, Managers, and Owners can create tasks");
 
             var priorityExists = await _context.TaskPriorities
@@ -172,7 +172,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group");
 
-            if (permissionLevel < (int)GroupRole.TeamLead && task.CreatedBy != userId)
+            if (permissionLevel < GroupRole.TeamLead && task.CreatedBy != userId)
                 return Error.Forbidden("Only the task creator or Team Leads and above can change task priority");
 
             if (!string.IsNullOrWhiteSpace(updateTaskDto.Title))
@@ -236,7 +236,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group");
 
-            var canChangeStatus = permissionLevel >= (int)GroupRole.TeamLead
+            var canChangeStatus = permissionLevel >= GroupRole.TeamLead
                 || task.AssignedToId == userId
                 || task.CreatedBy == userId;
 
@@ -296,7 +296,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group");
 
-            if (permissionLevel < (int)GroupRole.TeamLead)
+            if (permissionLevel < GroupRole.TeamLead)
                 return Error.Forbidden("Only Team Leads and above can change task priority");
 
             var oldPriority = task.Priority.Name;
@@ -342,7 +342,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group");
 
-            if (permissionLevel < (int)GroupRole.TeamLead)
+            if (permissionLevel < GroupRole.TeamLead)
                 return Error.Forbidden("Only Team Leads and above can assign tasks");
 
             var assigneeIsMember = await _context.GroupMembers
@@ -387,7 +387,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group");
 
-            var canUnassign = permissionLevel >= (int)GroupRole.TeamLead
+            var canUnassign = permissionLevel >= GroupRole.TeamLead
                 || task.AssignedToId == userId;
 
             if (!canUnassign)
@@ -418,7 +418,7 @@ namespace Plantitask.Infrastructure.Services
             if (permissionLevel == null)
                 return Error.Forbidden("You must be a member of this group");
 
-            if (permissionLevel < (int)GroupRole.Manager)
+            if (permissionLevel < GroupRole.Manager)
                 return Error.Forbidden("Only Managers and Owners can delete tasks");
 
             var now = DateTime.UtcNow;
